Phase2/backend-app/requirements.txt
fastapi[standard]>=0.100.0
uvicorn>=0.20.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9.0
PyJWT[crypto]>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
openai>=1.0.0
openai-agents>=0.0.1
mcp>=1.0.0